Measurement UI Creator | Address mypy and bandit issues (#28)
What does this Pull Request accomplish?
This pull request accomplishes the usage of release ni-measurement-plugin-sdk-service module. It also addresses the mypy and bandit issues.

Why should this Pull Request be merged?
As part of publishing the repository, the UI Creator is planned to be merged to the main branch. To do so, it is required to address the mypy issues, bandit issues and use the released version of the dependency module ni-measurement-plugin-sdk-service.

What testing has been done?
Manual testing.